refactored readme and config files naming. both autotasks working now
