+++ defender/auto-fund-chainlink-subscription/sample.config.yml
@@ -1,0 +1,14 @@
+# vrfCoordinatorV2 address on Goerli Testnet
+vrfCoordinatorAddress: "0x2Ca8E0C643bDe4C2E08ab1fA0da3401AdAD7734D"
+# LINK address on Goerli Testnet
+linkAddress: "0x326C977E6efc84E512bB9C30f76E30c160eD06FB"
+# WETH address on Goerli Testnet
+wethAddress: "0xB4FBF271143F4FBf7B91A5ded31805e42b2208d6"
+# Update with your Chainlink subscription id
+subscriptionId: 9893
+# Amount of LINK, if subscription balance falls below, will trigger automatically sending LINK
+threshold: "10"
+# Amount of LINK to add to your subscription
+fundAmount: "2"
+# Network to fund Chainlink subscription (mainnet or goerli)
+network: "goerli"
